+++ GraphIt.models/Edge.cs
@@ -17,7 +17,7 @@ namespace GraphIt.models
         [Required]
         public double Weight { get; set; }
         public string Label { get; set; }
-        public double Width { get; set; }
+        public int Width { get; set; }
         public bool Directed { get; set; }
         [Required]
         [RegularExpression(@"^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$")]
